Adding create member role during the installation
When I installed devstack,I encountered a
"No role with a name or ID of 'member'." error.
After reviewing the scripts of the installation
process,I found that during the deployment of
keystone,the creation of the member role was
really missing,so I modified it with reference
to the Queen version.

Change-Id: I5d589fd874a64fa9884e35cddc2b14a033fd1747