Update patch set 1
Patch Set 1:

(4 comments)

Hi Toni, welcome to DevStack and thanks for choosing to help to improve it. As you can see, we have a bit of a lack of reviewers, and you managed to submit your patch at the beginning of my holiday, so please excuse the lack of response so far.

The most important thing in order to proceed with this patch is to get the tests to pass, I added some comment regarding that below.

You may want to join the channel #openstack-qa on freenode for more direct communication, though that channel also isn't as busy as is used to be.

Patch-set: 1
Reviewer: Gerrit User 13252 <13252@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0