Make flavor_alt_ref default to m1.micro
When DEFAULT_INSTANCE_TYPE is not declared, the default flavor is
m1.nano and alt flavor is m1.micro. However, when it is defined, alt
flavor is chosen as the first available flavor (m1.tiny) which is not
enough when using advanced images like ubuntu or centos.

This patch defaults alt flavor to m1.micro and keeps the old logic for
whenever DEFAULT_INSTANCE_TYPE is m1.micro too as devstack ensures that
both flavors are different.

Change-Id: I0450fe2ab08bc103b1848bf9018468c1a8dc94e4
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>