Update patch set 4
Patch Set 4:

Reading Ian's comment I'm really not sure what to do with this. It's not about the wording for me, but adding the option to run devstack with apparmor enabled may generate a false sense of security that devstack doesn't really deliver.

Can we just disable apparmor for the OpenSuse jobs instead?

Patch-set: 4
Label: Workflow=0