Use openstack/placement instead of placement-in-nova
We introduce and set PLACEMENT_REPO, add a placement-manage command
to sync database tables (see one of the commits on which this
depends), use /etc/placement/placement.conf for config, and put the
uwsgi config file (pointing to placement-api instead of
nova-placement-api) in /etc/placement.

openstack/placement is also added to the required-projects in
the devstack-minimal zuul job.

Change-Id: I0b217e7a8c68a637b7a3445f6c44b7574117e320