Update patch set 8
Patch Set 8:

The remaining failures here are due to PROJECTS not being set right:

* http://logs.openstack.org/62/600162/7/check/legacy-devstack-dsvm-updown/0d805ef/logs/devstacklog.txt.gz#_2018-10-04_19_44_51_877
* http://logs.openstack.org/62/600162/7/check/neutron-tempest-linuxbridge/28c49ce/logs/devstacklog.txt.gz#_2018-10-04_19_45_57_619

The grenade depends-on is now passing. The database table create script needs some attention.

I'm away next week so if someone is able to pick this up and take it the rest of the way home, that would be fantastic.

Patch-set: 8