nova: Configure discover_hosts_in_cells_interval for multi-host
Deploy a multi-host configuration and try to boot an instance and you'll
be greeted with the following:

  Host 'foo' is not mapped to any cell.

The solution, as noted in the nova docs [1], is clear: either run
'nova-manage cell_v2 discover_hosts' on the controller or configure
'[scheduler] discover_hosts_in_cells_interval' so that nova will do this
for us. For the sanity of our users, let's do the latter.

[1] https://docs.openstack.org/nova/rocky/user/cells.html#fresh-install

Change-Id: I75f41d2633b751977b12eda6fccae9521dabd0be
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>