Update patch set 1
Patch Set 1:

> Why exactly?
 > 
 > This interval was basically a hack for ironic deployments that
 > didn't have discoverable nodes until after ironic resources were
 > enrolled and exposed. I think that now that we can discover by
 > service this is largely unnecessary in most cases. For devstack, it
 > would only be a thing for multinode setups, which are very much the
 > exception.
 > 
 > Is there some real situation that brought this up? Barring one, I
 > think I'm -1 on just enabling this "because."

As noted in the commit message

> Deploy a multi-host configuration and try to boot an instance and
> you'll be greeted with the following:
>
>   Host 'foo' is not mapped to any cell.
>
> The solution, as noted in the nova docs [1], is clear: either run
> 'nova-manage cell_v2 discover_hosts' on the controller or configure
> '[scheduler] discover_hosts_in_cells_interval' so that nova will do this
> for us. For the sanity of our users, let's do the latter.

So it's not "because" but rather because it seemed like the more automatic of the two methods of the the docs tell me I can and should use to resolve my issue. If this option is a hack then I can call out the need to run nova-manage in the multi-host doc instead, but the option should really be marked as advanced in nova, it should have it's hackiness mentioned as part of the option help text and that document should not tell people to use it. I can fix all of these but only if it's a hack (I don't know)

Patch-set: 1
Reviewer: Gerrit User 15334 <15334@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0