Update patch set 1
Patch Set 1:

It's definitely a hack, IMHO. We added it for ironic because they had a bunch of sequencing problems without it. Then with it, they still had problems because of the inherent raciness of needing to basically wait 2*interval before trying anything that assumed the hosts were up. After adding the by-service approach, they were finally able to have a repeatable process, and IIRC, tripleo removed the use of the periodic as a result.

Especially given the "why doesn't my ... oh, 30s later it does, hmm" behavior that people would experience, I'd much rather have it in the multinode docs explicitly.

Patch-set: 1