Support enable_plugin with TARGET_BRANCH
The existing function 'enable_plugin' clone the plugin
in master branch by default. This shouldn't be the default
behavior because the core projects will be cloned with
$TARGET_BRANCH which might be different from the plugin's branch.
This leads to an inconsistent of branches and break.

However, we don't want to modify the existing function as it
will break backward-compatibility. This commit introduces
a new function 'enable_plugin_in_current_branch' to address the issue.

Change-Id: I852eaa1ad33e3b8b8107f0d4c4950757383f53a0