Update patch set 2
Patch Set 2:

> Or do we simply need to get better at keeping this document in sync with what we check for in stackrc

We had a bit of debate about this in dib as well at some point, with some skew between what was in the documentation and what was available in reality.  If we're not testing it, it's almost certainly bitrotted in some way.  However, telling people in documentation "what's stable is what's being tested in the gate" is pretty unfriendly to someone new to the whole ecosystem, despite being technically true :)

As a bit of a middle-ground I ended up proposing something that automatically kept a table of what was being tested in the gate [1].  Now we have info like that in the local zuul config, maybe if someone was motivated we could have something similar.

https://review.openstack.org/#/c/418204/

Patch-set: 2