Update patch set 1
Patch Set 1: Code-Review-1

I think it's sound to expect that only requirements.txt is installed and not test-requirements.txt. Mostly it LGTM, but I do want to note that this could break some code that was depending on that. 

As such, 

(1) [reason of the -1] it should include a release note explaining the rationale and the potential impact of the change

(2) [non-voting] we should weigh if we want to make it not breaking in a first iteration by adding a parameter that defaults to install test-requirements.txt but that can be overridden. Then again, I don't think we should expect test-requirements.txt to be installed as part of pip_install and forcing it sounds like incorrect to me.

Patch-set: 1
Label: Code-Review=-1