Unset service type in systemd unit files when non-uwsgi
If a service is stacked to run with using UWSGI, a system unit file is
written with a Service Type of "notify". This means that the exec'd
process is expected to notify systemd when it is "ready". This works
nicely with uwsgi.

If, later, the same service is reconfigured on the same machine to not
use uwsgi, the service will timeout starting up because systemd is
waiting for a notification it will never get.

This change uses 'inidelete' to remote the service type if it is already
in the unit file. If it is not the command does nothing.

Closes-Bug: #1806023
Change-Id: Ib6c87b310113aca9d2f285020e1edd65174d35b6