Allow Glance to run non-uwsgi when other services do
Sometimes it is useful to be able to run Glance not using uwsgi while
also allowing other services to do so. As the code was written, if the
global setting for WSGI_MODE was 'uwsgi' then glance _had_ to run using
'uwsgi' and there was no other option.

This change adds GLANCE_WSGI_MODE which defaults to the value of
WSGI_MODE, but can be overriden in local.conf. If it is any value other
that 'uwsgi', the eventlet server will be used.

This can be important because sometimes glance (in particular, because
of image uploads) and uwsgi simply don't play well together and being
able to turn it off to determine if that is the issue is useful.

Change-Id: I62c5a52aa0d30bb02843c8d5abdf76669964671f