Update patch set 2
Patch Set 2:

> We've been back and forth with this before, at one point we had an
 > environment variable then it was all removed with 
 > I7c95927b9daad15766aac9d1276b10ca62efb24c.  I'd be against changing
 > the status quo of this now as I'm sure plenty of people depend on
 > these packages being installed.
 > 
 > If I'm understanding, in the linked bug you're using extras to
 > install ceilometer only in certain circumstances.  ISTM doing
 > something like
 > 
 > ceilometer[ceilometer] >= x.y.z
 > 
 > in your test-requirements file would work?

I am not sure what you mean. Our problem is with the ceilometer dependency and not with a specific version of ceilometer. Ceilometer is just a package which we would need in very specific deployments.

The problem we have now is that we want some n-odl zuul jobs to test with ceilometer and some others without ceilometer. To do so, we added ceilometer into test-requirements.txt and then, in our tox.ini, specified whether we were testing with requirements.txt (no ceilometer) or test-requirements.txt (with ceilometer). However, that logic is wrong as devstack is always using test-requirements.txt

Patch-set: 2