Update patch set 12
Patch Set 12:

Hmm, the error in the IBM CI is still there, not sure what is different there from how nova-manage is being installed. I'm also not sure whether we should consider this a blocker or whether we should put the burden onto the CI operators to fix their setups. I'll try to run the patch in a local environment and see what happens there.

http://sng01.objectstorage.softlayer.net/v1/AUTH_1940ea10-6e82-4501-b2f9-eb236510e575/ibmzkvmci/production/600162/12/check-tempest-dsvm-neutron-full-ubuntu-xenial-s390x/33739af/logs/devstacklog.txt.gz

Patch-set: 12