Add Fedora 29 support
Fedora 29 has just released and it had not support on devstack
installation, support were only for fedora 27 and 29.
Running the stack.sh script threw an error that it has not been
tested on fedora 29, and asked to run it with FORCE=yes.

This change add Fedora 29 as a supported distro version.

Change-Id: I29b688681b019a3240837b3b592d48da7a9353da