Drop user_name:neutron from neutron context_is_admin policy
DevStack defines a custom policy that adds a user 'neutron'
to admin role in neutron. When some other services need to
issue API operations to neutron with admin role, 'advsv' role
should be added to such service users.

This was added in commit 08a5fcc7faae8cab558617b46b684009df595fdd.
At that time, advsvc role support in neutron was not mature,
but it is well-supported now.
There is no reason to have a DevStack special policy now.

Change-Id: Iace703a780cc21c20907433180897311b0ba5c35