Remove spaces on variable NETWORK_API_EXTENSIONS
This allows users to provide the list of extensions with spaces
in between. For example:

  NETWORK_API_EXTENSIONS="ext1, ext2, ext3"

We will remove the spaces in between, so the result would be:

  "ext1,ext2,ext3"

This is because if the variable is provided by yaml (i.e. Zuul job)
and the list is too long so we want to break it into multiple lines.
Yaml will replace newline with space so we need to remove the space.

Needed-By: https://review.openstack.org/#/c/619642/

Change-Id: Ib61dadd305a7e105633761ac92f6fdefbe039459