Update patch set 1
Patch Set 1:

This is needed by https://review.openstack.org/#/c/513160/ to enable testing of this code in the nova-next job. Apparently turning this testing on just missed the Queens GA, but the dependent nova change (and feature) was implemented in Queens so it would be good to have test coverage for that code.

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0