Enable uwsgi debug logging
We're seeing slow nova-api start times in some CI
job runs, like an unaccounted for 26 seconds in some
cases. Devstack will timeout waiting for nova-api to
start after 60 seconds, so half of the time being
unaccounted for is a mystery and hopefully debug
logging in uwsgi will show us what is going on.

Change-Id: Ie854c3e260e9df213a667bde7a09397427bfaf3f
Related-Bug: #1807044