Update patch set 5
Patch Set 5:

My -1 is really just so that we keep the change consistent between the actual things going on and messages.  worlddump should work on fedora too.

TBH before voting positively on this personally I'd want to understand a bit better why some bridges aren't listed; I don't know the ins and outs of this but it seems weird the replacement for a deprecated command that seems pretty simple like "list" is direct consultation of /sys/ interfaces; are we sure it's not revealing some other sort of issue?

Patch-set: 5