Do not merge, checking etcd process priority
We see a lot of etcd timeouts in kuryr-kubernetes gates, where we use
the etcd3 as Kubernetes backend. To fight this this commit sets highest
IO priority and higher niceness for the etcd process. Hopefully this
will help.

As of now commit is only used to test patches on Kuryr-Kubernetes.

Change-Id: I3ae1d04a6d9bfc96512eca5ab73f00035a009119