Fix an issue in iniset function
Given the file to be configured, if user "stack" even doesn't have read access,
the result of configuration is not expected. iniset with "-sudo" option will
always create the section and the option which we want to configure for each
calling, no matter whether this section and this option exist in the file or
not. The root cause is the calling of grep and ini_has_option in iniset don't
use the "sudo" option.

Change-Id: I9d21322046b7be411c4c7c28fefc24894fa2e131
Signed-off-by: Yi Wang <yi.c.wang@intel.com>