Update patch set 1
Patch Set 1:

Hello Jens, thank you for your comment.

If the original is correct, then the definition is wrong.

The definition says: "It should generally be the last component of the git repo path (e.g., if the plugin’s repo is openstack/devstack-foo, then the name here should be “foo”)".

In this case the last component of the repo path is not "foo" but "devstack-foo".

Patch-set: 1
Reviewer: Gerrit User 23851 <23851@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0