zuul: add devstack-consumer job
In certain cases, you may want to get an OpenStack cloud using
DevStack but run tests against it using Ansible rather than using
tox.

This adds a job which runs devstack in `pre` in order to allow
consumers to manually define a `run` section with their own job
content to consume the cloud.

The difference from the existing `devstack-tox-functional-consumer`
is that this does not rely on tox being there.

Change-Id: Iebb14459a374928305c5d47ba43d92fea4c68ba8