Update patch set 1
Patch Set 1: Code-Review+1

> What kind of negative test would you expect to see? With the new
 > regex, there isn't much left that could fail. I updated the
 > existing test to cover the case that was failing, so reverting to
 > the old regex would fail that test.

I concur. It is more important to enable contributors to proceed forward instead of blocking this change. If the reviewer feels there is insufficient test coverage, then it may be a good item for them to follow-up on since the request for negative testing of the regex seems more of a nitpick.

Patch-set: 1
Reviewer: Gerrit User 11655 <11655@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1