Enable tenant networking type to be set
Ironic's CI makes extensive use of VLAN based networking
and the newer neutron plugin hardcodes the tenant networking
type to vxlan which is naturally problematic.

Enables the value to be set with a new parameter which is in-line
with the current plugin naming scheme, and supports the prior name
used in the neutron plugins and neutron-legacy plugin.

Change-Id: I55874c1ce82898e9dfb81505d8f3b14abde33579