Update patch set 1
Patch Set 1:

Hardlinking when cloning between trees on the same filesystem is Git's normal behavior, so we needed to work around that. I agree world-writeable directories are a security concern in many contexts, though I'm not sure that concern applies to test frameworks like devstack-gate. I'll be the first to admit there were likely other solutions to this particular problem, but as we didn't intend for anyone to keep using the legacy autotranslated roles long-term it served as an expedient means of moving on to the variety of other migration issues we were having to tackle at the same time.

This is >1year later. I had honestly hoped (and I think so did many of us) that the "legacy" roles, along with zuul-cloner and devstack-gate, would be dead and buried by now. Most of the resistance to changing established behavior here is 1. we can't know for sure who it might break and 2. it legitimizes/encourages continued use of this deprecated and basically frozen toolchain.

Patch-set: 1