Update patch set 1
Patch Set 1:

In the grenade jobs now, nova-cpu.conf is populated but the DB entries are removed as expected:

http://logs.openstack.org/11/631811/1/check/neutron-grenade/fba9842/logs/etc/nova/nova-cpu.conf.txt.gz

Looks like after grenade upgrades nova-compute and runs the new side devstack (master + this change) then it uses nova-cpu.conf:

http://logs.openstack.org/11/631811/1/check/neutron-grenade/fba9842/logs/screen-n-cpu.txt.gz#_Jan_18_17_46_31_099797

And the errors are gone.

The errors in the beginning of the log are from nova-compute running with nova.conf because of queens devstack not having this fix yet.

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0