Explictly start iscsid on bionic too for nova-compute
Recently iscsid was disabled by default on Ubuntu 18.04 (bionic),
and it may be on Xenial too, see:

https://bugs.launchpad.net/ubuntu/+source/open-iscsi/+bug/1755858

On a local Bionic deployment with Python 3, the lack of iscsid makes
nova-compute fail with an exception when trying to attach a volume:

Invalid input received: Connector doesn't have required information: initiator

The additional start should not hurt even on Xenial. Maybe it could be
enabled everywhere; even if I didn't find a reference for it, a second
execution of `systemctl start` may be idempotent (and probably with
service too).

This does not seem to be an issue on CentOS 7 (but Python 2) where
the socket activation of iscsid seems to work, so maybe there is
another way to make this working. Also, the service could be
enabled, not just started.

Change-Id: Ifa995dcf8eb930e959f54e96af6f5fce3eac28ae