Update patch set 3
Patch Set 3:

> I'm just wondering how this works in the gate bionic tests?
 > 
 > Possibly the difference there is that it is pre-installed during
 > image build [1] so it has been through a boot process?
 > 
 > But I checked on a fresh node and iscsid is not running just after
 > boot

That's a good question; I thought that it was enabled somehow on the infra nodes.

I agree that this should be catched by some test. In my case the image is not a cirros one, but a xenial image, but I suspect it should not matter. The other difference compared to the normal gate jobs is that my devstack configuration also includes devstack-plugin-ceph, even if it was configured just to replace swift through RGW, so again it should not matter.

Patch-set: 3
Reviewer: Gerrit User 10459 <10459@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0