Update patch set 3
Patch Set 3:

@Ghanshyam Mann

as i pointed out if you use a ubuntu cloud image that is
not built with nodepool builder then the
the tempest-full jobs will fail when it executes the cinder
senario tests.

the reason its passing in the gate
is likely becasue we install iscsi in the base element.

https://github.com/openstack/diskimage-builder/blob/master/diskimage_builder/elements/base/install.d/00-baseline-environment

Patch-set: 3