Move local.conf creation to run-devstack role
There may be situations where you'd like to include variables into
devstack's local.conf provided at runtime, rather than inserted
statically in the job: definition.

The first thing required facilitate this is to move writing of the
local.conf from the pre.yaml playbook to be the run-devstack role.
This gives users of the run-devstack role a chance to run things
before the local.conf is actually written out.

The first thought is that the devstack_localrc variable can then be
modified (have keys appended to, etc.) at runtime.  This raises a bit
of a chicken-egg problem with variables, however.  The "local_mtu" and
"external_bridge_mtu" variables might be used in a job's definition of
devstack_localrc which Zuul constructs.  These are "dangling" as it
were; for example a role can not append to the Zuul constructed
devstack_localrc dictionary because that involves expanding the
missing variables.  Even if you have at a random point:

 - debug:
    var: devstack_localrc

You'll get:

 VARIABLE IS NOT DEFINED!: 'external_bridge_mtu' is undefined

We could move the variable definitions into a separate role that users
of run-devstack would have to know to run *before* they try to modify
devstack_localrc, but this is getting needlessly complex.  Thus we
introduce a new argument "devstack_localrc_extra" to run-devstack to
pass an arbitrary dictionary of keys.

Thus the final situation to set a variable dynamically looks something
like:

 - import_role:
     name: foo
  # ^ does magic and sets variable foo_bar, which we want devstack to see

 - import_role
     name: run-devstack
   vars:
     devstack_localrc_extra:
       FOO: "\{\{ foo_bar \}\}"

Needed-By: https://review.openstack.org/634366
Change-Id: Id6731c5610d09d68e7d35f9b7b84909b2168ef26