Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

This is a lot of touch points for something I'm not sure we understand how it even happened.

PYTHON2_VERSION & PYTHON3_VERSION should be set by stackrc [1] and even if a job tried to set PTYHON3_VERSION='' my reading of the code is that the value would *still* be set.  If things are being sourced without stackrc then i imagine there are many other problems.

Since I can't see that these should ever be unset, wouldn't a better sanity check be a single check that the $1 value of get_pip_command was never unset?  It seems that would avoid the case of falling back to just calling "pip"

[1] http://git.openstack.org/cgit/openstack-dev/devstack/tree/stackrc#n139

Patch-set: 2
Label: Code-Review=-1