Update patch set 4
Patch Set 4: Code-Review+1

This is used as-is in just about every plugin ...

It might be neater to make a _v2 function, convert all in-tree users and then make the original into a wrapper that gives a deprecation notice and calls the _v2 function dropping the third argument.

To me this seems better than having IGNORED forever more across all the code, and gives an easy and grepable way to see what's using the old and new functions.

I did suggest this on the first go around of this where we first made the backwards incompatible change I2098d4eb2747282622cf486fa7dbf216f932f58b :)  Since this has become so widely used and cargo-cult copied, that's why I'm of the feeling that giving people at least a message that what they're doing is not working would be helpful.  If others disagree, however, I'm not that fussed.

Patch-set: 4
Label: Code-Review=+1