Update patch set 1
Patch Set 1:

OK I think the issue is that when running nova-metadata under uwsgi, we don't set the conductor indirection API which would otherwise happen if running under eventlet:

https://github.com/openstack/nova/blob/3cfcd117ce7ddb5a8d94396191cde559db73fb86/nova/cmd/api_metadata.py#L45

We should likely be doing that here:

https://github.com/openstack/nova/blob/3cfcd117ce7ddb5a8d94396191cde559db73fb86/nova/api/metadata/wsgi.py#L19

So this is a nova bug. Jens - I'll open a nova bug and propose a patch which hopefully you can test out in your environment since it appears we don't have any multinode jobs that run with the separate n-api-meta service?

Patch-set: 1