Update patch set 2
Patch Set 2: Code-Review-1

Queens uses the UCA that was available while it was being developed, which is Pike. Changing this retroactively may cause side-effects in other projects, so we cannot easily do this. If there is a specific issue you want to fix with this, please create a bug report so we can look at it in more detail.

Patch-set: 2
Reviewer: Gerrit User 13252 <13252@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1