Remove the usage of read_password from library files
The read_password function is defined inside stack.sh
and it cannot be used inside the "public library interface"
provided by DevStack.

While the call to read_password may be moved to stack.sh,
both requirements are base ones and an error when they are
not specified does not look like a significant issue.

Generally speaking, read_password seems a relic from the past
and devstack does not seem to be advertised as an interactive
tool.

Change-Id: I8adc6723b677dfac2bef735f660e056c498bf773