Update patch set 5
Patch Set 5:

@frickler: Apologies, I missed your first comment, didn't mean to ignore it.

Cinderlib is tightly coupled with Cinder and we wanted to have them both in the same repository [1][2][3], but since we wanted to have 2 different PyPi packages it had to go on a different repository.

The idea of installing cinderlib whenever we install Cinder was to preserve this tight coupling and facilitate running the cinderlib functional tests in all Cinder and 3rd party Cinder CIs, as well as testing by developers that deploy Devstack without the need of extra steps.

If you don't think that's enough to submit the patch to devstack I'll go with the plugin route.

[1]: https://review.openstack.org/#/c/620669/
[2]: https://review.openstack.org/#/c/620670/
[3]: https://review.openstack.org/#/c/620671/

Patch-set: 5