Use split instead of strip
If pip major version is greater than 9 stack.sh fails due
to issue with incorrect version. strip return 1, split
will return 10 or higher

It was in past in review:

Properly get pip version

The old code was strip()ing the version string instead of split()ing the
version string so we always got the first character of the version
string. This worked fine as long as the pip version was single digit but
as soon as it rolls over to '10.stuff' we will compare:

  pip version 1 (instead of 10) > 6

Which fails bceause 1 is less than six. Instaed we really do want to
compare 10 > 6 so use split on '.' instead.

Closes-Bug: #1764046
Change-Id: Ic7d0c04d7fa77774ab2d70fb9d11f182becec553
(cherry picked from commit 0657795)i

Change-Id: Ie810b22eacd81c3f648f351933106b3397c69773