Add option for glance wsgi mode
Currently glance use wsgi_mode to control to use standalone mode or
in uwsgi mode which is global option that controls other service as
well.
Adding option for glance so that it can choose wsgi mode independantly.
Its default value is set to uwsgi for compatibility, but user should
favor standalone mode according to suggestion:
https://docs.openstack.org/glance/latest/admin/apache-httpd.html

Change-Id: I1f89a220c31c8b51a2276aacb2cf8084d6c313e8