Debug check_python3_support_for_package_remote
We are hitting issues with parsing the python 3 classifier
from the pypi metadata json API response in rax-dfw for
some reason. One theory is that we are getting an empty
or partial result, so this patch adds debug logging if we
hit either case. It also fixes a logging statement associated
with the result of check_python3_support_for_package_remote
to use the package name rather than the package directory name
since we are checking pypi for the former.

Change-Id: I5a8003106a5a10a9e0d9d9d964f7a3858a765b8d
Related-Bug: #1820892