Update patch set 6
Patch Set 6:

@Ian: So in order to avoid wrongly interpreting this, do you object to using py36 and would like to see py35 instead? Or are you claiming that centos7 folks will continue to use py27 even when it is EOLd upstream and we should cater for that?

The issue is that there seems to be no easy way to remove the "pip<10" cap running with py27, because there are lots of dependencies in various system packages. Nobody has volunteered to work on that, either.

We could decide to keep the cap only for centos7, but that would make that job a special snowflake in a different way, which I'm not sure is the better solution.

We did keep the pip cap in place for almost a year now and I think that was way too long, so in the interest of devstack as a whole, I'd rather drop centos7 support than keeping the cap for another cycle.

Patch-set: 6
Label: Verified=0