Update patch set 3
Patch Set 3:

Oh, right.

So to actually prove this bug, we would need something like multiple regions, and show a case where the ironicclient picks an endpoint in a region *other* than the one specified in the conf.

I'm not real savvy on devstack; how hard would it be to set that up? (And is it something we could make permanent?)

Patch-set: 3