Update patch set 3
Patch Set 3:

> Oh, right.
 > 
 > So to actually prove this bug, we would need something like
 > multiple regions, and show a case where the ironicclient picks an
 > endpoint in a region *other* than the one specified in the conf.
 > 
 > I'm not real savvy on devstack; how hard would it be to set that
 > up? (And is it something we could make permanent?)

Yea - that would be ideal, but short term this will at least make sure that the code path acts the same way when a region is provided.

Testing multi regions gets a bit tricky, as even if you have two regions, it's still possible that it will pick the "right" region. Since there is no "default" region, so the result is random(ish).

In an ideal scenario we would have a test with two regions, and attempt to connect to both regions (to make sure we are not just "accidentally" connecting to the expected region).

Patch-set: 3
Reviewer: Gerrit User 22623 <22623@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0